# contracts/openapi.yaml
openapi: 3.0.4
info:
  title: Todo API
  version: 1.0.0
paths:
...
  /api/settings:
    get:
      operationId: getSettings
      tags:
        - Settings
      responses:
        '200':
          description: The stored settings.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SettingsResponse'
    put:
      operationId: updateSettings
      tags:
        - Settings
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SettingsRequest'
      responses:
        '200':
          description: The stored settings.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SettingsResponse'
        '400':
          description: The request is not valid.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
  /api/settings/jira-token:
    put:
      operationId: setJiraToken
      tags:
        - Settings
      summary: Stores the Jira personal access token the import authenticates with.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/JiraTokenRequest'
      responses:
        '200':
          description: The token was stored.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SettingsResponse'
        '400':
          description: The token was empty.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
    delete:
      operationId: clearJiraToken
      tags:
        - Settings
      summary: Forgets the stored Jira token.
      responses:
        '200':
          description: The token was removed.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SettingsResponse'
  /api/settings/ado-token:
    put:
      operationId: setAdoToken
      tags:
        - Settings
      summary: Stores the Azure DevOps personal access token the import authenticates with.
      description: >-
        A route of its own rather than a field on PUT /api/settings, because that is a full
        replacement which reads an absent field as "clear": a client saving anything else on
        the settings page would wipe the token, and the token can never be sent back to the
        client to be echoed. SettingsResponse carries only hasAdoToken.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AdoTokenRequest'
      responses:
        '200':
          description: The token was stored.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SettingsResponse'
        '400':
          description: The token was empty.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
    delete:
      operationId: clearAdoToken
      tags:
        - Settings
      summary: Forgets the stored Azure DevOps token.
      responses:
        '200':
          description: The token was removed.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SettingsResponse'
  /api/settings/autostart:
    put:
      operationId: enableAutostart
      tags:
        - Settings
      summary: Starts the app when the user signs in to Windows.
      description: >-
        A route of its own for the same reason the tokens have one: PUT /api/settings is a full
        replacement that reads an absent field as "clear", so saving a language would switch
        autostart off. PUT enables and DELETE disables, rather than a body carrying a boolean,
        because the two verbs already say it.

        Unlike every other setting this one is not stored in the database. It lives under
        HKCU\Software\Microsoft\Windows\CurrentVersion\Run, which is where Windows actually reads
        it from - so that key is the truth, and a copy in the Settings table could only disagree
        with it. Remove the key with another tool and the switch reports off, which is correct.
      responses:
        '200':
          description: Autostart is on.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SettingsResponse'
        '400':
          description: Autostart could not be turned on.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
    delete:
      operationId: disableAutostart
      tags:
        - Settings
      summary: Stops the app from starting when the user signs in.
      responses:
        '200':
          description: Autostart is off.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SettingsResponse'
        '400':
          description: Autostart could not be turned off.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
...
components:
  schemas:
    ApiError:
      type: object
      additionalProperties: false
      required: [code, message]
      properties:
        code: { type: string }
        message: { type: string }
...
    SettingsRequest:
      type: object
      additionalProperties: false
      properties:
        language:
          type: string
          nullable: true
          example: da
        delegates:
          type: array
          items:
            type: string
          description: >-
            People you hand tasks to, offered as suggestions when a task moves to WaitingFor. A
            suggestion list, not a closed set: the who field stays free text, because waiting on
            somebody unlisted — or on nobody at all — are both valid states.
        jiraBaseUrl:
          type: string
          nullable: true
        jiraProjectKey:
          type: string
          nullable: true
        jiraWaitingStatuses:
          type: array
          items:
            type: string
        jiraIncludeWaiting:
          type: boolean
        jiraDoneStatuses:
          type: array
          items:
            type: string
          description: >-
            Statuses that mean the issue is finished. No switch beside it, unlike the waiting and duty
            pairs: doneness only ever offers to close a task you already have, so there is nothing to
            opt into. An empty list is valid and means no suggestions.
        jiraDutyStatuses:
          type: array
          items:
            type: string
          description: >-
            Statuses that mean "waiting for the shared duty pool". Issues in these are fetched
            regardless of assignee when jiraOnDuty is on, and they arrive actionable rather than
            waiting.
        jiraOnDuty:
          type: boolean
          description: >-
            Whether the user currently holds the 2nd level support duty. Off by default. Separate
            from the list so the list survives going off duty.
        adoBaseUrl:
          type: string
          nullable: true
          description: The Azure DevOps collection URL, which may contain a space in its name.
        adoProject:
          type: string
          nullable: true
        adoDoneStates:
          type: array
          items:
            type: string
          description: >-
            States that mean the work item is finished. The counterpart of jiraDoneStatuses, and its
            own list rather than a shared one: the two systems spell different words.
        adoWaitingStates:
          type: array
          items:
            type: string
          description: >-
            States that mean "waiting for somebody else". Compared ordinally: two states that
            differ only in case are two states the system keeps apart.
        adoIncludeWaiting:
          type: boolean
          description: Whether rows in a waiting state are imported anyway. Off by default.
        adoWorkItemTypes:
          type: array
          items:
            type: string
          description: >-
            The work item types to import, so test artefacts stay out. A requirement rather than
            an optional filter: an empty list is rejected with ado.workItemTypesRequired, and an
            absent one restores the default Bug, User Story, Task. The plan said an empty list
            meant every type; that was wrong twice over. It is the same trap as the empty project
            key — the absence of a limit is not a neutral default — and the storage cannot carry
            the claim anyway, because an empty list is stored as no row at all and the reader
            cannot tell never-configured from deliberately-emptied. Compared ordinally, like
            adoWaitingStates.
        adoDefaultDeadlineDays:
          type: integer
          format: int32
          default: 3
          description: >-
            How many days ahead the import sets a deadline, because Azure DevOps has no due date
            field of its own. 0 means no deadline. Not nullable: a nullable field would add an
            @if branch in the frontend, and 0 reads as "off" for a number of days. Rejected
            below 0 or above 365 with ado.defaultDeadlineDaysInvalid.

            The default is load-bearing here and only here. This is a full replacement, so an
            absent field means clear — but System.Text.Json gives 0 for an absent int, and 0 is
            a meaningful value, so without the default a request that never mentioned this field
            would silently turn the deadline off. NSwag turns the default into a property
            initializer, which the deserialiser leaves alone for an absent field, so absent binds
            to 3 and a deliberate 0 stays 0. SettingsResponse deliberately carries no default:
            an initializer there would let a handler that never assigned the field answer 3 and
            make the test for it unfailable.
    SettingsResponse:
      type: object
      additionalProperties: false
      required:
        - delegates
        - jiraWaitingStatuses
        - jiraIncludeWaiting
        - jiraDutyStatuses
        - jiraOnDuty
        - jiraDoneStatuses
        - hasJiraToken
        - adoWaitingStates
        - adoDoneStates
        - adoIncludeWaiting
        - adoWorkItemTypes
        - adoDefaultDeadlineDays
        - hasAdoToken
        - autostart
      properties:
        language:
          type: string
          nullable: true
          example: da
        delegates:
          type: array
          items:
            type: string
          description: >-
            People you hand tasks to, offered as suggestions when a task moves to WaitingFor. A
            suggestion list, not a closed set: the who field stays free text, because waiting on
            somebody unlisted — or on nobody at all — are both valid states.
        jiraBaseUrl:
          type: string
          nullable: true
        jiraProjectKey:
          type: string
          nullable: true
        jiraWaitingStatuses:
          type: array
          items:
            type: string
        jiraIncludeWaiting:
          type: boolean
        jiraDoneStatuses:
          type: array
          items:
            type: string
          description: >-
            Statuses that mean the issue is finished. No switch beside it, unlike the waiting and duty
            pairs: doneness only ever offers to close a task you already have, so there is nothing to
            opt into. An empty list is valid and means no suggestions.
        jiraDutyStatuses:
          type: array
          items:
            type: string
          description: >-
            Statuses that mean "waiting for the shared duty pool". Issues in these are fetched
            regardless of assignee when jiraOnDuty is on, and they arrive actionable rather than
            waiting.
        jiraOnDuty:
          type: boolean
          description: >-
            Whether the user currently holds the 2nd level support duty. Off by default. Separate
            from the list so the list survives going off duty.
        hasJiraToken:
          type: boolean
          description: >-
            Whether a token is stored. The token itself is never returned; it is written
            through PUT /api/settings/jira-token and cleared through DELETE.
        adoBaseUrl:
          type: string
          nullable: true
          description: The Azure DevOps collection URL, which may contain a space in its name.
        adoProject:
          type: string
          nullable: true
        adoDoneStates:
          type: array
          items:
            type: string
          description: >-
            States that mean the work item is finished. The counterpart of jiraDoneStatuses, and its
            own list rather than a shared one: the two systems spell different words.
        adoWaitingStates:
          type: array
          items:
            type: string
          description: >-
            States that mean "waiting for somebody else". Compared ordinally: two states that
            differ only in case are two states the system keeps apart.
        adoIncludeWaiting:
          type: boolean
          description: Whether rows in a waiting state are imported anyway. Off by default.
        adoWorkItemTypes:
          type: array
          items:
            type: string
          description: >-
            The work item types to import, so test artefacts stay out. Never empty: an absent row
            reads back as the default Bug, User Story, Task, and PUT /api/settings rejects an
            empty list rather than storing one. See the request schema for why.
        adoDefaultDeadlineDays:
          type: integer
          format: int32
          description: >-
            How many days ahead the import sets a deadline, because Azure DevOps has no due date
            field of its own. 0 means no deadline. Always the effective value: an absent row
            reads back as 3. No default is declared here on purpose — see the request schema.
        hasAdoToken:
          type: boolean
          description: >-
            Whether a token is stored. The token itself is never returned; it is written
            through PUT /api/settings/ado-token and cleared through DELETE.
        autostart:
          type: boolean
          description: >-
            Whether the app starts when the user signs in to Windows. Read from the registry
            rather than from the Settings table, because the registry is what Windows reads it
            from - so it is the only answer that cannot be stale. Absent from SettingsRequest on
            purpose: it is written through PUT and DELETE /api/settings/autostart.
    JiraTokenRequest:
      type: object
      additionalProperties: false
      required: [token]
      properties:
        token:
          type: string
    AdoTokenRequest:
      type: object
      additionalProperties: false
      required: [token]
      properties:
        token:
          type: string
...